refactoring

+++ lib/cli/oacis_cli_analysis.rb
@@ -17,9 +17,7 @@ class OacisCli < Thor
     anz_parameters = Hash[mapped]
 
     return if options[:dry_run]
-    unless options[:yes]
-      return unless overwrite_file?(options[:output])
-    end
+    return unless options[:yes] or overwrite_file?(options[:output])
     File.open(options[:output], 'w') do |io|
       # for visibility, manually print the json object as follows
       io.puts "[", "  #{anz_parameters.to_json}", "]"
@@ -108,9 +106,7 @@ class OacisCli < Thor
 
   ensure
     return if options[:dry_run]
-    unless options[:yes]
-      return unless overwrite_file?(options[:output])
-    end
+    return unless options[:yes] or overwrite_file?(options[:output])
     write_analysis_ids_to_file(options[:output], analyses)
   end
