+++ lib/cli/oacis_cli_parameter_set.rb
@@ -17,9 +17,7 @@ class OacisCli < Thor
     parameter_set = Hash[mapped]
 
     return if options[:dry_run]
-    unless options[:yes]
-      return unless overwrite_file?(options[:output])
-    end
+    return unless options[:yes] or overwrite_file?(options[:output])
     File.open(options[:output], 'w') do |io|
       # for visibility, manually print the json object as follows
       io.puts "[", "  #{parameter_set.to_json}", "]"
@@ -73,9 +71,7 @@ class OacisCli < Thor
 
   ensure
     return if options[:dry_run]
-    unless options[:yes]
-      return unless overwrite_file?(options[:output])
-    end
+    return unless options[:yes] or overwrite_file?(options[:output])
     write_parameter_set_ids_to_file(options[:output], parameter_sets)
   end
 
@@ -94,3 +90,4 @@ class OacisCli < Thor
     end
   end
 end
+
